Fix
git-svn-id: svn+ssh://svn.yandex.ru/yt/trunk@907 41d65440-b5be-11dd-afe3-b2e846d9b4f8

+++ yt/ytlib/bus/packet.cpp
@@ -12,7 +12,7 @@ static NLog::TLogger& Logger = BusLogger;
 const int THeaderTraits<TPacketHeader>::FixedSize =
     sizeof (TPacketHeader);
 const int THeaderTraits<TMultipartPacketHeader>::FixedSize =
-    (int) &(((TMultipartPacketHeader*) NULL)->PartSizes);
+    (size_t) &(((TMultipartPacketHeader*) NULL)->PartSizes);
 
 ////////////////////////////////////////////////////////////////////////////////
